add build for tokenizer
